+++ viteclient/src/App.jsx
@@ -3,13 +3,13 @@ import { useCookies } from 'react-cookie';
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import LoginForm from './LoginForm';
-import HomePage from './HomePage';
+import HomePage from './pages/VillaAnna/HomePage';
 import BewaesserungPage from './BewaesserungPage';
 import MarkisePage from './MarkisePage';
 import SettingsPage from './SettingsPage';
 import AuthGuard from './AuthGuard';
 import axios from 'axios';
-import NotFoundPage from './404Page';
+import NotFoundPage from './pages/404Page';
 
 function App() {
   const [cookies] = useCookies(['session']);